requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-phonenumber-field==5.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
lxml==4.6.3
phonenumbers==8.12.22
phonenumberslite==8.12.22
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
